+++ tests/testthat/test_spatial.R
@@ -8,6 +8,10 @@ yws$crop_factor = NA
 yws$crop_factor[yws$land_cover_type=="agriculture"] = 0.75
 
 data("examplemeteo")
+examplemeteo2 <- examplemeteo
+examplemeteo2$dates <- as.Date(row.names(examplemeteo2))
+row.names(examplemeteo2) <- NULL
+
 data("SpParamsMED")
 dates = seq(as.Date("2001-03-01"), as.Date("2001-03-03"), by="day")
 
@@ -22,22 +26,38 @@ test_that("Can simulate three days over landscape",{
                              SpParams = SpParamsMED, progress = FALSE), "sf")
 })
 
+test_that("Can simulate three days over landscape with dates as column",{
+  expect_s3_class(spwb_spatial(ypts, meteo = examplemeteo2, dates = dates, 
+                               SpParams = SpParamsMED, progress = FALSE), "sf")
+  expect_s3_class(spwb_spatial(yws, meteo = examplemeteo2, dates = dates, 
+                               SpParams = SpParamsMED, progress = FALSE), "sf")
+  expect_s3_class(growth_spatial(ypts, meteo = examplemeteo2, dates = dates, 
+                                 SpParams = SpParamsMED, progress = FALSE), "sf")
+  expect_s3_class(growth_spatial(yws, meteo = examplemeteo2, dates = dates, 
+                                 SpParams = SpParamsMED, progress = FALSE), "sf")
+})
 
-test_that("Can simulate one year with management",{
+
+test_that("Can simulate one year forest dynamics with management",{
   expect_s3_class(fordyn_spatial(ypts[1,], meteo = examplemeteo, 
                                SpParams = SpParamsMED, progress = FALSE), "sf")
 })
 
+test_that("Can simulate one year forest dynamics with management with dates in column",{
+  expect_s3_class(fordyn_spatial(ypts[1,], meteo = examplemeteo2, 
+                                 SpParams = SpParamsMED, progress = FALSE), "sf")
+})
 
-# test_that("Can simulate three days over landscape using old interpolator",{
-#   data("exampleinterpolationdata")
-#   expect_s3_class(spwb_spatial(ypts[1,], meteo = exampleinterpolationdata, 
-#                                SpParams = SpParamsMED, progress = FALSE), "sf")
-# })
-# 
-# test_that("Can simulate three days over landscape using new interpolator",{
-#   interpolator = meteoland::with_meteo(meteoland_meteo_example) |>
-#     meteoland::create_meteo_interpolator(params = defaultInterpolationParams(), verbose = FALSE)
-#   expect_s3_class(spwb_spatial(ypts[1,], meteo = interpolator,
-#                                SpParams = SpParamsMED, progress = FALSE), "sf")
-# })
+test_that("Can simulate 3 days over landscape using new meteoland interpolator",{
+  interpolator = meteoland::with_meteo(meteoland_meteo_example, verbose = FALSE) |>
+    meteoland::create_meteo_interpolator(params = defaultInterpolationParams(), verbose = FALSE)
+  datesMeteo <- as.Date(stars::st_get_dimension_values(interpolator, "date"))
+  expect_s3_class(spwb_spatial(ypts[1,], meteo = interpolator, dates = datesMeteo[1:3],
+                               SpParams = SpParamsMED, progress = FALSE), "sf")
+  expect_s3_class(spwb_spatial(yws, meteo = interpolator, dates = datesMeteo[1:3],
+                               SpParams = SpParamsMED, progress = FALSE), "sf")
+  expect_s3_class(growth_spatial(ypts[1,], meteo = interpolator, dates = datesMeteo[1:3],
+                               SpParams = SpParamsMED, progress = FALSE), "sf")
+  expect_s3_class(spwb_spatial(yws, meteo = interpolator, dates = datesMeteo[1:3],
+                               SpParams = SpParamsMED, progress = FALSE), "sf")
+})