+++ NEWS.md
@@ -6,6 +6,7 @@
 - IMPORTANT: In this version we removed sp-type S4 classes that were deprecated in 1.0.1. Use release version 1.2.1 to continue using deprecated classes.
 - New function 'fordyn_land'
 - Compliance with medfate 3.1.1
+- Can accept a list of interpolator objects as weather input
 
 # Version 1.2.1
 - Compliance with medfate 3.0.0